+++ src/index.js
@@ -10,3 +10,11 @@ export { startServer } from './server.js';
 
 // Installer exports
 export { install } from './install.js';
+
+// Startup helper exports
+export {
+  ensureServerRunning,
+  renderOnboardingWalkthrough,
+  shouldShowOnboarding,
+  markOnboardingComplete
+} from './start.js';